% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{acf_no_center}
\alias{acf_no_center}
\title{Sample Covariance Function Without Centering}
\usage{
acf_no_center(data, lag = 0)
}
\arguments{
\item{data}{A numeric matrix.}

\item{lag}{An integer specifying the lag.}
}
\value{
A numeric matrix representing the covariance matrix.
}
\description{
This function computes the sample covariance matrix with an optional lag.
}
