% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{bn.factor.number}
\alias{bn.factor.number}
\title{Factor number estimator of Bai and Ng (2002)}
\usage{
bn.factor.number(x, covx = NULL, q.max = NULL, do.plot = FALSE, center = TRUE)
}
\arguments{
\item{x}{input time series matrix, with each row representing a variable}

\item{covx}{covariance of \code{x}}

\item{q.max}{maximum number of factors; if \code{q.max = NULL}, a default value is selected as \code{min(50, floor(sqrt(min(dim(x)[2] - 1, dim(x)[1]))))}}

\item{do.plot}{whether to plot the value of the information criterion}

\item{center}{whether to de-mean the input \code{x} row-wise}
}
\value{
a list containing
\item{q.hat}{ the mimimiser of the chosen information criteria}
\item{sv}{ svd of \code{covx}}
}
\description{
Estimates the number of restricted factors by minimising an information criterion.
Currently three information criteria proposed in Owens, Cho, and Barigozzi (2022) (\code{pen.op = 1,2,3}) are implemented,
with \code{pen.op = 2} recommended as a default choice based on numerical experiments.
}
\details{
See Bai and Ng (2002) for further details.
}
\examples{
library(fnets)

set.seed(123)
n <- 500
p <- 50
common <- sim.restricted(n, p)
idio <- sim.var(n, p)
x <- common$data * apply(idio$data, 1, sd) / apply(common$data, 1, sd) + idio$data

bn <- factor.number(x, restricted = TRUE, do.plot = TRUE)
bn$q.hat
}
\references{
preprint

Bai, J. & Ng, S. (2002) Determining the number of factors in approximate factor models. Econometrica. 70: 191-221. \cr

Alessi, L., Barigozzi, M.,  and Capasso, M. (2010) Improved penalization for determining the number of factors in approximate factor models. Statistics & Probability Letters, 80(23-24):1806–1813.
}
\keyword{internal}
