% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{cross_val}
\alias{cross_val}
\title{internal function that carries out cross validation to choose tuning parameter for data truncation}
\usage{
cross_val(data, n_tau = 60, lag, standardise = TRUE)
}
\description{
internal function that carries out cross validation to choose tuning parameter for data truncation
}
\keyword{internal}
