% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{cross_val_lag}
\alias{cross_val_lag}
\title{internal function that carries out cross validation to choose tuning parameter for data truncation}
\usage{
cross_val_lag(data, n_tau, lag, standardise = standardise)
}
\description{
internal function that carries out cross validation to choose tuning parameter for data truncation
}
\keyword{internal}
