% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{cv_trunc}
\alias{cv_trunc}
\title{Truncate data, with truncation parameter chosen by cross-validation.}
\usage{
cv_trunc(data, n_tau = 60, lag = 0, cv_lag = 0, standardise = T)
}
\arguments{
\item{data}{Input time series each column representing a time series variable.}

\item{n_tau}{An integer that determines the number of taus to use in grid used for cross-validation}

\item{lag}{This is an integer argument that is used when the \code{cv_trunc} function is used for (auto)covariance estimation, of particular lag.
The lag determines which (auto)covariance matrix is used in tuning.}

\item{cv_lag}{An integer argument, that is used when the \code{cv_trunc} function is used with data modeled as a factor-adjusted VAR.
The integer determines up to what lag auto-covariance matrix is used in the cv-measure.
 In implementation, this will be set as default to be the VAR order. When \code{cv_lag = 0}, only the (auto)covariance matrix, of lag determined
 by the \code{lag} argument, is used.}

\item{standardise}{boolean; whether to scale up the truncation parameter for each series by the MAD of the corresponding series.}
}
\description{
Truncate data, with truncation parameter chosen by cross-validation.
}
