% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{direct.cv}
\alias{direct.cv}
\title{Cross-validation for the constrained l1-minimisation problem for inverse matrix estimation}
\usage{
direct.cv(
  object,
  xx,
  target = c("spec", "acv"),
  symmetric = c("min", "max", "avg", "none"),
  path.length = 10,
  n.folds = 1,
  q = 0,
  kern.bandwidth.const = 4,
  n.cores = min(parallel::detectCores() - 1, 3)
)
}
\description{
Cross-validation for the constrained l1-minimisation problem for inverse matrix estimation
}
\keyword{internal}
