% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{dyn.pca}
\alias{dyn.pca}
\title{Dynamic PCA}
\usage{
dyn.pca(xx, q = NULL, ic.op = 4, kern.const = 4)
}
\arguments{
\item{xx}{centred input time series matrix, with each row representing a variable}

\item{q}{number of factors. If \code{q = NULL}, the factor number is estimated by an information criterion-based approach of Hallin and Liška (2007)}

\item{ic.op}{choice of the information criterion. Currently the three options from Hallin and Liška (2007) (\code{ic.op = 1, 2} or \code{3}) and
their variations with logarithm taken on the cost (\code{ic.op = 4, 5} or \code{6}) are implemented,
with \code{ic.op = 5} recommended as a default choice based on numerical experiments}

\item{kern.const}{constant multiplied to \code{floor((dim(x)[2]/log(dim(x)[2]))^(1/3)))} which determines the kernel bandwidth for dynamic PCA}
}
\value{
a list containing
\item{q}{ number of factors}
\item{hl}{ if \code{q = NULL}, the output from \link[fnets]{hl.factor.number}}
\item{spec}{ a list containing the estimates of the spectral density matrices for \code{x}, common and idiosyncratic components}
\item{acv}{ a list containing estimates of the autocovariance matrices for \code{x}, common and idiosyncratic components}
\item{kern.const}{ input parameter}
}
\description{
Performs principal components analysis in frequency domain for identifying common and idiosyncratic components.
}
\keyword{internal}
