% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{factor.number}
\alias{factor.number}
\title{Factor number estimators of Hallin and Liška (2007) and Bai and Ng (2002)}
\usage{
factor.number(
  x,
  restricted = FALSE,
  q.max = NULL,
  mm = NULL,
  w = NULL,
  covx = NULL,
  do.plot = FALSE,
  center = TRUE
)
}
\arguments{
\item{x}{input time series matrix, with each row representing a variable}

\item{restricted}{whether to estimate the number of restricted or unrestricted factors}

\item{q.max}{maximum number of factors; if \code{q.max = NULL}, a default value is selected as \code{min(50, floor(sqrt(min(dim(x)[2] - 1, dim(x)[1]))))}}

\item{mm}{bandwidth; defaults to \code{floor(4 * (dim(x)[2]/log(dim(x)[2]))^(1/3)))}}

\item{w}{vector of length \code{2 * mm + 1} containing symmetric weights; if \code{w = NULL}, default weights are generated using the Bartlett kernel and \code{mm}}

\item{covx}{covariance matrix of x}

\item{do.plot}{whether to plot the information criteria values}

\item{center}{whether to de-mean the input \code{x} row-wise}
}
\value{
a list containing
\item{q.hat}{ a vector containing minimisers of the six information criteria}
\item{sv}{ a list containing the singular value decomposition of \code{Sigma_x}}
and if \code{restricted = FALSE}
\item{Gamma_x}{ an array containing the estimates of the autocovariance matrices of \code{x} at \code{2 * mm + 1} lags}
\item{Sigma_x}{ an array containing the estimates of the spectral density matrices of \code{x} at \code{2 * mm + 1} Fourier frequencies}
}
\description{
Estimates the number of factors by minimising an information criterion over sub-samples of the data.
For \code{restricted = FALSE}, the three information criteria proposed in Hallin and Liška (2007) (\code{pen.op = 1, 2} or \code{3})
and their variations with logarithm taken on the cost (\code{pen.op = 4, 5} or \code{6}) are implemented,
with \code{pen.op = 5} recommended as a default choice based on numerical experiments.
For \code{restricted = TRUE}, the three information criteria in Owens, Cho, and Barigozzi (2022) are implemented,
with \code{pen.op = 2} recommended by default.
}
\details{
For further details, see Hallin and Liška (2007), Bai and Ng (2002), or Alessi, Barigozzi, and Capasso (2010).
}
\examples{
library(fnets)

set.seed(123)
n <- 500
p <- 50
common <- sim.unrestricted(n, p)
idio <- sim.var(n, p)
x <- common$data * apply(idio$data, 1, sd) / apply(common$data, 1, sd) + idio$data

hl <- factor.number(x, restricted = FALSE, do.plot = TRUE)
hl$q.hat
library(fnets)

set.seed(123)
n <- 500
p <- 50
common <- sim.restricted(n, p)
idio <- sim.var(n, p)
x <- common$data * apply(idio$data, 1, sd) / apply(common$data, 1, sd) + idio$data

bn <- factor.number(x, restricted = TRUE, do.plot = TRUE)
bn$q.hat
}
\references{
Hallin, M. & Liška, R. (2007) Determining the number of factors in the general dynamic factor model. Journal of the American Statistical Association, 102(478), 603--617.

Bai, J. & Ng, S. (2002) Determining the number of factors in approximate factor models. Econometrica. 70: 191-221. \cr

Alessi, L., Barigozzi, M., and Capasso, M. (2010) Improved penalization for determining the number of factors in approximate factor models. Statistics & Probability Letters, 80(23-24):1806–1813.
}
