% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{fit.var}
\alias{fit.var}
\title{Regularised Yule-Walker estimation for VAR processes}
\usage{
fit.var(
  x,
  lambda,
  symmetric = "min",
  idio.var.order = 1,
  idio.method = c("ds", "lasso"),
  n.cores = min(parallel::detectCores() - 1, 3),
  niter = 100,
  tol = 0,
  do.plot = FALSE,
  center = TRUE
)
}
\arguments{
\item{x}{input time series matrix, with each row representing a time series}

\item{lambda}{regularisation parameter}

\item{symmetric}{type of symmetry to enforce on \code{Gamma}, one of 'min', 'max', 'avg', 'none'}

\item{idio.var.order}{order of idiosyncratic VAR model}

\item{idio.method}{A string specifying the type of l1-regularised estimator for the idiosyncratic VAR matrix, possible values are:
\itemize{
   \item{\code{'lasso'}}{ Lasso estimator}
   \item{\code{'ds'}}{ Dantzig Selector}
}}

\item{n.cores}{number of cores to use for parallel computing (\code{'ds'} only)}

\item{niter}{maximum number of descent steps (\code{'lasso'} only)}

\item{tol}{numerical tolerance for increases in the loss function(\code{'lasso'} only)}

\item{do.plot}{return a plot of the loss function against descent steps (\code{'lasso'} only)}

\item{center}{demean the input \code{x}}
}
\value{
A list which contains the following fields:
\itemize{
\item{beta: VAR parameters}
\item{lambda: regularisation parameter}
\item{Gamma: Estimated noise covariance}
}
}
\description{
Returns parameter estimates for the idiosyncratic VAR and the corresponding Gamma matrix, either by the Dantzig selector or Lasso methods
}
\details{
Further information can be found in Barigozzi, Cho and Owens (2021).
}
\examples{
require(fnets)
require(doParallel)
require(lpSolve)
set.seed(222)
n <- 200
p<- 100
chi <- sim.factor.M2(n,p)
xi <- sim.idio(n,p)
sample.data <- chi$data + xi$data
fit.var(sample.data, .1, idio.method = "lasso")
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
