% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{fnets.var.internal}
\alias{fnets.var.internal}
\title{internal function for \code{fnets.var}}
\usage{
fnets.var.internal(
  xx,
  acv,
  method = c("lasso", "ds"),
  lambda = NULL,
  var.order = 1,
  tuning.args = list(tuning = c("cv", "bic"), n.folds = 1, penalty = NULL, path.length
    = 10),
  do.threshold = FALSE,
  n.iter = NULL,
  tol = 0,
  n.cores = min(parallel::detectCores() - 1, 3)
)
}
\description{
internal function for \code{fnets.var}
}
\keyword{internal}
