% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{idio.cv}
\alias{idio.cv}
\title{Cross-validation for l1-regularised VAR estimation}
\usage{
idio.cv(
  xx,
  lambda.max = NULL,
  var.order = 1,
  idio.method = c("lasso", "ds"),
  path.length = 10,
  n.folds = 1,
  q = 0,
  kern.bandwidth.const = 4,
  cv.plot = TRUE
)
}
\arguments{
\item{xx}{centred input time series matrix, with each row representing a time series}

\item{lambda.max}{maximum regularisation parameter, if NULL this is set to the smallest which sets all entries to 0}

\item{var.order}{vector of VAR orders to consider}

\item{idio.method}{estimation method, one of "lasso" or "ds"}

\item{path.length}{number of regularisation parameters to consider}

\item{n.folds}{number of CV folds}

\item{q}{factor number}

\item{kern.bandwidth.const}{constant to determine bandwidth size}

\item{cv.plot}{return a plot of the CV error against regularisation parameters, stratified by VAR order}
}
\value{
A list which contains the following fields:
\itemize{
\item{lambda: minimising argument}
\item{var.order: minimising order}
\item{cv.error: matrix of errors}
\item{lambda.path: candidate lambda values}
}
}
\description{
Cross-validation for l1-regularised VAR estimation
}
\details{
Further information can be found in Barigozzi, Cho and Owens (2021).
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
