% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{idio.predict}
\alias{idio.predict}
\title{Prediction for the idiosyncratic VAR process}
\usage{
idio.predict(object, x, cpre, h = 1)
}
\arguments{
\item{object}{fnets object}

\item{x}{input time series matrix, with each row representing a time series}

\item{cpre}{estimated common component}

\item{h}{forecast horizon}
}
\value{
A list containing
\itemize{
\item{is: in-sample estimation}
\item{fc: forecast}
\item{h: forecast horizon}
}
}
\description{
Predictsidiosyncratic components from a fnets object for new data
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
