% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{make.gg}
\alias{make.gg}
\title{Construct \code{gg} and \code{GG} matrices for Yule-Walker estimation}
\usage{
make.gg(acv, d)
}
\arguments{
\item{acv}{autocovariance array}

\item{d}{order of VAR}
}
\description{
Construct \code{gg} and \code{GG} matrices for Yule-Walker estimation
}
\keyword{internal}
