% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{network}
\alias{network}
\title{Convert networks into igraph objects}
\usage{
network(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{additional arguments}
}
\description{
Convert networks into igraph objects
}
\seealso{
\link[fnets]{network.fnets}
}
