% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{nonpar.lrpc}
\alias{nonpar.lrpc}
\title{Nonparametric partial coherence matrix estimation}
\usage{
nonpar.lrpc(
  object,
  x,
  eta = NULL,
  symmetric = c("min", "max", "avg", "none"),
  lrpc.cv.args = list(n.folds = 1, path.length = 10),
  n.cores = min(parallel::detectCores() - 1, 3)
)
}
\arguments{
\item{object}{fnets object}

\item{x}{input time series matrix, with each row representing a time series}

\item{eta}{regularisation parameter, if NULL this selected by cross-validation}

\item{symmetric}{type of symmetry to enforce on output, one of 'min', 'max', 'avg', 'none'}

\item{lrpc.cv.args}{A list specifying arguments to the cross-validation (CV) procedure containing:
\itemize{
   \item{\code{'n.folds'}}{ number of folds}
   \item{\code{'path.length'}}{ number of lambda values to consider}
}}

\item{n.cores}{number of cores to use for parallel computing}
}
\value{
A list containing
\itemize{
\item{\code{'Omega'}}{ Estimated partial coherence matrix}
\item{\code{'eta'}}{ regularisation parameter}
}
}
\description{
Returns a non-parametric estimate of the partial coherence matrix, possibly using cross-validation
}
\examples{
#nonpar.lrpc
require(doParallel)
require(lpSolve)
model <- fnets(sample.data, q=2, idio.method = "lasso")
nonpar.lrpc(model, sample.data, 1, n.cores = 1)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
