% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{par.lrpc}
\alias{par.lrpc}
\title{Parametric estimation of long-run partial correlations of factor-adjusted VAR processes}
\usage{
par.lrpc(
  object,
  x,
  eta = NULL,
  cv.args = list(n.folds = 1, path.length = 10, do.plot = FALSE),
  do.correct = TRUE,
  n.cores = min(parallel::detectCores() - 1, 3)
)
}
\arguments{
\item{object}{\code{fnets} object}

\item{x}{input time series matrix; with each row representing a variable}

\item{eta}{regularisation parameter; if \code{eta = NULL}, it is selected by cross validation}

\item{cv.args}{a list specifying arguments for the cross validation procedure
for selecting the tuning parameter involved in long-run partial correlation matrix estimation. It contains:
\itemize{
   \item{\code{n.folds}}{ number of folds}
   \item{\code{path.length}}{ number of regularisation parameter values to consider; a sequence is generated automatically based in this value}
   \item{\code{do.plot}}{ whether to plot the output of the cross validation step}
}}

\item{do.correct}{whether to correct for any negative entries in the diagonals of the inverse of long-run covariance matrix}

\item{n.cores}{number of cores to use for parallel computing, see \link[parallel]{makePSOCKcluster}}
}
\value{
a list containing
\item{Delta}{ estimated inverse of the innovation covariance matrix}
\item{Omega}{ estimated inverse of the long-run covariance matrix}
\item{pc}{ estimated innovation partial correlation matrix}
\item{lrpc}{ estimated long-run partial correlation matrix}
\item{eta}{ regularisation parameter}
}
\description{
Returns a parametric estimate of long-run partial correlations of the VAR process
from the VAR parameter estimates and the inverse of innovation covariance matrix obtained via constrained \code{l1}-minimisation.
}
\details{
See Barigozzi, Cho and Owens (2021) for further details.
}
\examples{
\dontrun{
set.seed(123)
n <- 500
p <- 50
common <- sim.common1(n, p)
idio <- sim.var(n, p)
x <- common$data + idio$data
out <- fnets(x, q = NULL, idio.method = 'lasso', lrpc.method = 'none')
plrpc <- par.lrpc(out, x, cv.args = list(n.folds = 1, path.length = 10, do.plot = TRUE))
out$lrpc <- plrpc
out$lrpc.method <- 'par'
plot(out, type = 'pc', display = 'network', threshold = .05)
plot(out, type = 'lrpc', display = 'heatmap', threshold = .05)
}
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2021) FNETS: Factor-adjusted network analysis for high-dimensional time series.
}
