% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{plot.fnets.pcn}
\alias{plot.fnets.pcn}
\title{Plot fnets.pcn object}
\usage{
\method{plot}{fnets.pcn}(object, names = NULL, groups = NULL, threshold = 0, size = NULL, ...)
}
\description{
Plot fnets.pcn object
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
