% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{plot_internal}
\alias{plot_internal}
\title{internal function for \code{plot.fnets} and \code{network}}
\usage{
plot_internal(
  x,
  type = c("granger", "pc", "lrpc"),
  display = c("network", "heatmap"),
  names = NA,
  groups = NA,
  group.colours = NA,
  ...
)
}
\description{
internal function for \code{plot.fnets} and \code{network}
}
\keyword{internal}
