% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{predict.fm}
\alias{predict.fm}
\title{Forecasting for factor models}
\usage{
\method{predict}{fm}(object, x, h = 1, forecast.restricted = TRUE, r = c("ic", "er"), ...)
}
\arguments{
\item{object}{\code{fm} object}

\item{x}{input time series matrix, with each row representing a variable}

\item{h}{forecasting horizon}

\item{forecast.restricted}{whether to forecast using a restricted or unrestricted, blockwise VAR representation of the common component}

\item{r}{number of restricted factors, or a string specifying the factor number selection method when \code{forecast.restricted = TRUE};
 possible values are:
\itemize{
   \item{\code{"ic"}}{ information criteria of Bai and Ng (2002)}
   \item{\code{"er"}}{ eigenvalue ratio}
}}

\item{...}{not used}
}
\value{
a list containing
\item{is}{ in-sample predictions}
\item{forecast}{ forecasts for the given forecasting horizon}
\item{r}{ factor number}
}
\description{
Produces forecasts of the data for a given forecasting horizon by
estimating the best linear predictors of the common component
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2021) Factor-adjusted network estimation and forecasting for high-dimensional time series. arXiv preprint arXiv:2201.06110.

Ahn, S. C. & Horenstein, A. R. (2013) Eigenvalue ratio test for the number of factors. Econometrica, 81(3), 1203--1227.
}
\seealso{
\link[fnets]{fnets.factor.model}, \link[fnets]{common.predict}
}
