% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{print.fm}
\alias{print.fm}
\title{Print factor model}
\usage{
\method{print}{fm}(x, ...)
}
\arguments{
\item{x}{\code{fm} object}

\item{...}{not used}
}
\value{
NULL, printed to console
}
\description{
Prints a summary of a \code{fm} object
}
\examples{
set.seed(123)
n <- 500
p <- 50
common <- sim.restricted(n, p)
idio <- sim.var(n, p)
x <- common$data + idio$data
out <- fnets.factor.model(x, q = 2)
print(out)
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2022) FNETS: Factor-adjusted network estimation and forecasting for high-dimensional time series. arXiv preprint arXiv:2201.06110.

Owens, D., Cho, H. & Barigozzi, M. (2022) fnets: An R Package for Network Estimation and Forecasting via Factor-Adjusted VAR Modelling
}
\seealso{
\link[fnets]{fnets.factor.model}
}
