% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{print.threshold}
\alias{print.threshold}
\title{Print threshold}
\usage{
\method{print}{threshold}(x, ...)
}
\arguments{
\item{x}{\code{threshold} object}

\item{...}{not used}
}
\value{
NULL, printed to console
}
\description{
Prints a summary of a \code{threshold} object
}
\examples{
set.seed(123)
A <- diag(.7, 50) + rnorm(50^2, 0, .1)
threshold.A <- threshold(A)
print(threshold.A)
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2022) FNETS: Factor-adjusted network estimation and forecasting for high-dimensional time series. arXiv preprint arXiv:2201.06110.

Owens, D., Cho, H. & Barigozzi, M. (2022) fnets: An R Package for Network Estimation and Forecasting via Factor-Adjusted VAR Modelling
}
\seealso{
\link[fnets]{threshold}
}
