% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.data}
\alias{sample.data}
\title{Simulated example common+idiosyncratic time series}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 200 columns.
}
\usage{
sample.data
}
\description{
Example `p=100`, `n=200` time series, where the common component is generated using Model 1 and the idiosyncratic component is generated with a sparse VAR.
}
\keyword{datasets}
