% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.common2}
\alias{sim.common2}
\title{Simulate data from a static factor model}
\usage{
sim.common2(n, p, q = 2)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{q}{number of dynamic factors; number of static factors is given by \code{2 * q}}
}
\value{
a list containing
\item{data}{ generated series}
\item{q}{ number of factors}
\item{r}{ number of static factors}
}
\description{
Simulate the common component following a dynamic factor model that admits a static representation;
see the model (C2) in the reference.
}
\examples{
common <- sim.common2(500, 50)
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2021) FNETS: Factor-adjusted network analysis for high-dimensional time series.
}
