% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.factor.M1}
\alias{sim.factor.M1}
\title{Simulate data from a dynamic factor model (Model 1)}
\usage{
sim.factor.M1(
  n,
  p,
  trunc.lags = 20,
  do.scale = T,
  a1 = NULL,
  a2 = NULL,
  alpha1 = NULL,
  alpha2 = NULL
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of series}

\item{trunc.lags}{lag for moving average representation}

\item{do.scale}{scale the output (default \code{TRUE} )}

\item{a1, a2, alpha1, alpha2}{generative parameters (default null, see reference)}
}
\value{
A list containing
\itemize{
\item{\code{'data'}}{ generated series}
\item{\code{'shocks'}}{ 2-dimensional shock series}
\item{\code{'a1','a2','alpha1','alpha2'}}{ generative parameters}
}
}
\description{
Simulate data from a dynamic factor model (Model 1)
}
\examples{
    sim.factor.M1(100,10)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
\seealso{
\link{sim.factor.M2}
}
