% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.factor.M2}
\alias{sim.factor.M2}
\title{Simulate data from a static factor model (Model 2) with factor number \code{r = q * lags}}
\usage{
sim.factor.M2(n, p, q = 2, lags = 2, do.scale = T, loadings = NULL, D = NULL)
}
\arguments{
\item{n}{sample size}

\item{p}{number of series}

\item{q}{dynamic dimension (default 2)}

\item{lags}{number of lags for which the observed series depends on the factor series (default 2)}

\item{do.scale}{scale the output (default \code{TRUE} )}

\item{loadings}{loading matrix, dimension p by r (default null)}

\item{D}{transition matrix, dimension q by q (default null)}
}
\value{
A list containing
\itemize{
\item{\code{'data'}}{ generated series}
\item{\code{'shocks'}}{ \code{q}-dimensional shock series}
\item{\code{'factors'}}{ \code{q}-dimensional factor series}
\item{\code{'D'}}{ transition matrix}
\item{\code{'loadings'}}{ factor loadings}
}
}
\description{
Simulate data from a static factor model (Model 2) with factor number \code{r = q * lags}
}
\examples{
    sim.factor.M2(100,10)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
\seealso{
\link{sim.factor.M1}
}
