% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.idio}
\alias{sim.idio}
\title{Simulate data from a (sparse) VAR(1) model}
\usage{
sim.idio(
  n,
  p,
  A = NULL,
  cov = diag(1, p),
  prob = 1/p,
  two.norm = NULL,
  do.scale = T
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of series}

\item{A}{transition matrix, dimension p by p (default null)}

\item{cov}{generative covariance matrix (default identity)}

\item{prob}{probability of an edge existing in the transition matrix, if \code{A} is \code{NULL}}

\item{two.norm}{target 2-norm to scale A by (default NULL)}

\item{do.scale}{scale the output (default \code{TRUE})}
}
\value{
A list containing
\itemize{
\item{\code{'data'}}{ generated series}
\item{\code{'A'}}{ transition matrix}
}
}
\description{
Simulate data from a (sparse) VAR(1) model
}
\examples{
    sim.idio(100,10, A=diag(0.3, 10))
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
