% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.static}
\alias{sim.static}
\title{Simulate data from a static factor model}
\usage{
sim.static(n, p, q = 2, heavy = FALSE)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{q}{number of dynamic factors; number of static factors is given by \code{2 * q}}

\item{heavy}{if \code{heavy = FALSE}, common shocks are generated from \code{rnorm} whereas if \code{heavy = TRUE}, from \code{rt} with \code{df = 5} and then scaled by \code{sqrt(3 / 5)}}
}
\value{
a list containing
\item{data}{ generated series}
\item{q}{ number of factors}
\item{r}{ number of static factors}
}
\description{
Simulate the common component following a dynamic factor model that admits a static representation;
see the model (C2) in the reference.
}
\examples{
common <- sim.static(500, 50)
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2021) FNETS: Factor-adjusted network analysis for high-dimensional time series.
}
