% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.var}
\alias{sim.var}
\title{Simulate a VAR(1) process}
\usage{
sim.var(n, p, Gamma = diag(1, p))
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{Gamma}{innovation covariance matrix}
}
\value{
a list containing
\item{data}{ generated series}
\item{A}{ transition matrix}
\item{Gamma}{ innovation covariance matrix}
}
\description{
Simulate a VAR(1) process; see the reference for the generation of the transition matrix.
}
\examples{
idio <- sim.var(500, 50)
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2021) FNETS: Factor-adjusted network analysis for high-dimensional time series.
}
