% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{static.pca}
\alias{static.pca}
\title{Static PCA}
\usage{
static.pca(
  xx,
  q = NULL,
  q.method = c("ic", "er"),
  q.max = NULL,
  ic.op = 2,
  mm = NULL
)
}
\description{
Static PCA
}
\examples{
{
set.seed(123)
n <- 500
p <- 50
common <- sim.restricted(n, p)
idio <- sim.var(n, p)
x <- common$data + idio$data
fnets:::static.pca(x)
}
}
\keyword{internal}
