% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{tau_grid_stand_fun}
\alias{tau_grid_stand_fun}
\title{internal function for \code{cv_trunc}}
\usage{
tau_grid_stand_fun(data, n_steps, standardise)
}
\description{
internal function for \code{cv_trunc}
}
\keyword{internal}
