% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{truncateAndComputeCovariance_lag}
\alias{truncateAndComputeCovariance_lag}
\title{Sample Covariance Function Without Centering, after truncating data.}
\usage{
truncateAndComputeCovariance_lag(data, tau, lag = 0)
}
\arguments{
\item{data}{A numeric matrix.}

\item{tau}{A numeric vector.}

\item{lag}{An integer specifying the lag.}
}
\value{
A numeric matrix representing the covariance matrix.
}
\description{
This function truncated the data and computes the sample covariance matrix with an optional lag.
}
