% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{tuning_plot}
\alias{tuning_plot}
\title{Plotting output for tuning parameter selection in fnets}
\usage{
tuning_plot(x, ...)
}
\arguments{
\item{x}{\code{fnets} object}

\item{...}{additional arguments}
}
\value{
CV/IC plot for the VAR component, and CV plot for the lrpc component (when \code{x$do.lrpc = TRUE}).
}
\description{
Tuning plots for S3 objects of class \code{fnets}.
Produces up to two plots visualising CV and IC procedures for selecting tuning parameters and the VAR order.
}
\details{
See Barigozzi, Cho and Owens (2022) for further details.
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2022) FNETS: Factor-adjusted network estimation and forecasting for high-dimensional time series. arXiv preprint arXiv:2201.06110.

Owens, D., Cho, H. & Barigozzi, M. (2022) fnets: An R Package for Network Estimation and Forecasting via Factor-Adjusted VAR Modelling. arXiv preprint arXiv:2301.11675.
}
\seealso{
\link[fnets]{fnets}
}
\keyword{internal}
