% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unrestricted}
\alias{unrestricted}
\title{Simulated data from the unrestricted factor-adjusted vector autoregression model}
\format{
## `unrestricted`
A data frame with 500 rows (observations) and 50 columns (series)
}
\usage{
unrestricted
}
\description{
set.seed(123)
n <- 500
p <- 50
common <- sim.unrestricted(n, p)
idio <- sim.var(n, p)
x <- common$data + idio$data
}
\keyword{datasets}
