% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{var.lasso}
\alias{var.lasso}
\title{Lasso-type Yule-Walker estimation for VAR processes}
\usage{
var.lasso(
  GG,
  gg,
  lambda,
  symmetric = "min",
  niter = 100,
  tol = 0,
  do.plot = FALSE
)
}
\arguments{
\item{GG, gg}{output from \code{make.gg}}

\item{lambda}{regularisation parameter}

\item{symmetric}{type of symmetry to enforce on \code{Gamma}, one of 'min', 'max', 'avg', 'none'}

\item{niter}{maximum number of descent steps}

\item{tol}{numerical tolerance for increases in the loss function}

\item{do.plot}{return a plot of the loss function against descent steps}
}
\value{
A list which contains the following fields:
\itemize{
\item{\code{beta}}{VAR parameters}
\item{\code{lambda}}{ regularisation parameter}
\item{\code{Gamma}}{ Estimated noise covariance}
\item{\code{loss}}{ Objective function value}
}
}
\description{
Returns parameter estimates for the idiosyncratic VAR and the corresponding Gamma matrix
}
\details{
Further information can be found in Barigozzi, Cho and Owens (2021).
}
\examples{
require(fnets)
require(doParallel)
require(lpSolve)
dpca <- dyn.pca(sample.data, q=2)
mg <- make.gg(dpca$acv$Gamma_i,1)
vl <- var.lasso(mg$GG,mg$gg, .1)
vd <- var.dantzig(mg$GG,mg$gg, .1, n.cores = 1)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
