% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{yw.ic}
\alias{yw.ic}
\title{Information criterion for factor-adjusted VAR estimation}
\usage{
yw.ic(
  xx,
  method = c("lasso", "ds"),
  lambda.max = NULL,
  var.order = 1,
  penalty = NULL,
  path.length = 10,
  q = 0,
  kern.bw = NULL,
  do.plot = FALSE
)
}
\description{
Information criterion for factor-adjusted VAR estimation
}
\keyword{internal}
