% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{bn.factor.number}
\alias{bn.factor.number}
\title{Factor number estimator of Bai and Ng (2002)}
\usage{
bn.factor.number(
  x,
  lam = NULL,
  f = NULL,
  q.max = NULL,
  ic.op = 2,
  do.plot = FALSE,
  center = TRUE
)
}
\arguments{
\item{x}{input time series matrix, with each row representing a variable}

\item{lam, f}{loading and factor matrices; if \code{lam = NULL} or \code{f = NULL}, these are obtained with PCA}

\item{q.max}{maximum number of factors; if \code{q.max = NULL}, a default value is selected as \code{min(50, floor(sqrt(min(dim(x)[2] - 1, dim(x)[1]))))}}

\item{ic.op}{chosen information criterion}

\item{do.plot}{whether to plot the value of the information criterion}

\item{center}{whether to de-mean the input \code{x} row-wise}
}
\value{
a list containing
\item{q.hat}{ the mimimiser of the chosen information criteria}
\item{lam}{ loading matrix}
\item{f}{ factor series}
\item{q.max}{ maximum number of factors}
\item{ic}{ vector of information criteria values}
}
\description{
Estimates the number of restricted factors by minimising an information criterion.
Currently the five information criteria proposed in Bai and Ng (2002) (\code{ic.op = 1,...,5}) are implemented,
with \code{ic.op = 2} recommended as a default choice based on numerical experiments.
}
\details{
See Bai and Ng (2002) for further details.
}
\examples{
library(fnets)

set.seed(123)
n <- 500
p <- 50
common <- sim.restricted(n, p)
idio <- sim.var(n, p)
x <- common$data * apply(idio$data, 1, sd) / apply(common$data, 1, sd) + idio$data

bn <- bn.factor.number(x, q.max = NULL, center = FALSE, do.plot = TRUE)
bn$q.hat
}
\references{
Bai, J. & Ng, S. (2002) Determining the number of factors in approximate factor models. Econometrica. 70: 191-221. \cr
}
