% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common.irf.estimation}
\alias{common.irf.estimation}
\title{Blockwise VAR estimation under GDFM}
\usage{
common.irf.estimation(
  xx,
  Gamma_c,
  q,
  factor.var.order = NULL,
  max.var.order = NULL,
  trunc.lags,
  n.perm
)
}
\description{
Blockwise VAR estimation under GDFM
}
\keyword{internal}
