% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common.irf.estimation}
\alias{common.irf.estimation}
\title{Blockwise VAR estimation under GDFM}
\usage{
common.irf.estimation(
  xx,
  Gamma_c,
  q,
  factor.var.order = NULL,
  max.var.order = NULL,
  trunc.lags,
  n.perm
)
}
\description{
Blockwise VAR estimation under GDFM
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2022) FNETS: Factor-adjusted network estimation and forecasting for high-dimensional time series. arXiv preprint arXiv:2201.06110.

Forni, M., Hallin, M., Lippi, M. & Zaffaroni, P. (2017). Dynamic factor models with infinite-dimensional factor space: Asymptotic analysis. Journal of Econometrics, 199(1), 74--92.

Owens, D., Cho, H. & Barigozzi, M. (2022) fnets: An R Package for Network Estimation and Forecasting via Factor-Adjusted VAR Modelling
}
\keyword{internal}
