% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common.predict}
\alias{common.predict}
\title{Forecasting the factor-driven common component}
\usage{
common.predict(
  object,
  x,
  h = 1,
  common.method = c("restricted", "unrestricted"),
  r = NULL
)
}
\arguments{
\item{object}{\code{fnets} object}

\item{x}{input time series matrix, with each row representing a variable}

\item{h}{forecasting horizon}

\item{common.method}{a string specifying the method for common component forecasting; possible values are:
\itemize{
   \item{\code{"restricted"}}{ performs forecasting under a restrictive static factor model}
   \item{\code{"unrestricted"}}{ performs forecasting under an unrestrictive, blockwise VAR representation of the common component}
}}

\item{r}{number of static factors; if \code{common.method = "restricted"} and \code{r = NULL},
it is estimated as the maximiser of the ratio of the successive eigenvalues
of the estimate of the common component covariance matrix, see Ahn and Horenstein (2013)}
}
\value{
a list containing
\item{is}{ in-sample estimator of the common component}
\item{fc}{ forecasts of the common component for a given forecasting horizon \code{h}}
\item{r}{ static factor number}
\item{h}{ forecast horizon}
}
\description{
Produces forecasts of the common component
for a given forecasting horizon by estimating the best linear predictors
}
\examples{
set.seed(123)
n <- 500
p <- 50
common <- sim.common1(n, p)
idio <- sim.var(n, p)
x <- common$data + idio$data
out <- fnets(x, q = NULL, idio.var.order = 1, idio.method = "lasso", lrpc.method = "none")
cpre <- common.predict(out, x, h = 1, common.method = "restricted", r = NULL)
ipre <- idio.predict(out, x, cpre, h = 1)
}
\references{
Barigozzi, M., Cho, H. & Owens, D. (2021) FNETS: Factor-adjusted network analysis for high-dimensional time series. arXiv preprint arXiv:2201.06110.

Ahn, S. C. & Horenstein, A. R. (2013) Eigenvalue ratio test for the number of factors. Econometrica, 81(3), 1203--1227.

Forni, M., Hallin, M., Lippi, M. & Reichlin, L. (2005). The generalized dynamic factor model: one-sided estimation and forecasting. Journal of the American Statistical Association, 100(471), 830--840.

Forni, M., Hallin, M., Lippi, M. & Zaffaroni, P. (2017). Dynamic factor models with infinite-dimensional factor space: Asymptotic analysis. Journal of Econometrics, 199(1), 74--92.
}
