% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common.predict}
\alias{common.predict}
\title{Prediction for the factor-driven common component}
\usage{
common.predict(object, x, h = 1, common.method = c("static", "var"), r = NULL)
}
\arguments{
\item{object}{\code{fnets} object}

\item{x}{input time series matrix, with each row representing a time series}

\item{h}{forecast horizon}

\item{common.method}{which of "static" or "var" to forecast the common component with}

\item{r}{factor number, if \code{r = NULL} this is selected using the maximal eigenratio}
}
\value{
A list containing
\itemize{
\item{is}{in-sample estimator of the common component}
\item{fc}{forecasts of the common component for a given forecasting horizon \code{h}}
\item{r}{factor number}
\item{h}{forecast horizon}
}
}
\description{
Predicts common component from a \code{fnets} object for new data
}
\examples{
require(fnets)
set.seed(222)
n <- 200
p<- 100
chi <- sim.factor.M2(n,p)
xi <- sim.idio(n,p)
sample.data <- chi$data + xi$data
model <- fnets(sample.data, q=2, idio.method = "lasso")
pr <- predict(model,sample.data, common.method = "static")
cpre <- common.predict(model,sample.data, common.method = "static")
ip <- idio.predict(model,sample.data, cpre)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.

Forni, M., Hallin, M., Lippi, M., & Reichlin, L. (2005). The generalized dynamic factor model: one-sided estimation and forecasting. Journal of the American Statistical Association, 100(471), 830--840.

Forni, M., Hallin, M., Lippi, M., & Zaffaroni, P. (2017). Dynamic factor models with infinite-dimensional factor space: Asymptotic analysis. Journal of Econometrics, 199(1), 74--92.
}
