% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.restricted}
\alias{data.restricted}
\title{Simulated data from the restricted factor-adjusted vector autoregression model}
\format{
## `data.restricted`
A ts object with 500 rows (observations) and 50 columns (series)
}
\usage{
data.restricted
}
\description{
\code{set.seed(123)} \cr
\code{n <- 500} \cr
\code{p <- 50} \cr
\code{common <- sim.restricted(n, p)} \cr
\code{idio <- sim.var(n, p)} \cr
\code{x <- common$data + idio$data}
}
\keyword{datasets}
