% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{dyn.pca}
\alias{dyn.pca}
\title{Dynamic PCA}
\usage{
dyn.pca(xx, q = NULL, ic.op = 4, kern.bandwidth.const = 4)
}
\arguments{
\item{xx}{centred input time series matrix, with each row representing a time series}

\item{q}{the number of factors, if q=NULL this is selected by the information criterion-based estimator of Hallin and Liska (2007)}

\item{ic.op}{an index number for the information criterion}

\item{kern.bandwidth.const}{constant to determine bandwidth size}
}
\value{
A list containing
\itemize{
\item{q: see Arguments}
\item{spec: Spectral density matrices}
\item{acv: Autocovariance matrices}
\item{kern.bandwidth.const: Constant to determine bandwidth size}
}
}
\description{
Performs principal components analysis of the autocovariance matrices.
}
\examples{
dyn.pca(sample.data, q=2)
}
