% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{dyn.pca}
\alias{dyn.pca}
\title{Dynamic PCA}
\usage{
dyn.pca(xx, q = NULL, ic.op = 4, kern.const = 4)
}
\arguments{
\item{xx}{centred input time series matrix, with each row representing a time series}

\item{q}{the number of factors, if q=NULL this is selected by the information criterion-based estimator of Hallin and Liska (2007)}

\item{ic.op}{an index number for the information criterion (1 to 6)}

\item{kern.const}{constant to determine bandwidth size}
}
\value{
A list containing
\itemize{
\item{\code{'q'}}{ number of factors}
\item{\code{'spec'}}{ Spectral density matrices}
\item{\code{'acv'}}{ Autocovariance matrices}
\item{\code{'kern.const'}}{ Constant to determine bandwidth size}
}
}
\description{
Performs principal components analysis of the autocovariance matrices.
}
