% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{fnets}
\alias{fnets}
\title{Factor-adjusted network analysis}
\usage{
fnets(
  x,
  q = NULL,
  ic.op = 4,
  kern.bandwidth.const = 4,
  common.var.args = list(var.order = 1, max.var.order = NULL, trunc.lags = 20, n.perm =
    10),
  idio.var.order = 1,
  idio.method = c("ds", "lasso"),
  idio.cv.args = list(n.folds = 1, path.length = 10, symmetric = "min", cv.plot = TRUE),
  center = TRUE
)
}
\arguments{
\item{x}{input time series matrix, with each row representing a time series}

\item{q}{the number of factors, if q=NULL this is selected by the information criterion-based estimator of Hallin and Liska (2007)}

\item{ic.op}{an index number for the information criterion}

\item{kern.bandwidth.const}{constant to determine bandwidth size}

\item{common.var.args}{A list specifying the estimator for the common component. This contains:
\itemize{
   \item{\code{'var.order'}}{the order of the VAR model, if NULL then selected blockwise by BIC}
   \item{\code{'max.var.order'}}{the maximum order of the VAR model for the BIC to consider}
   \item{\code{'trunc.lags'}}{the order of the MA representation}
   \item{\code{'n.perm'}}{number of cross-sectional permutations}
}}

\item{idio.var.order}{order of idiosyncratic VAR model}

\item{idio.method}{A string specifying the type of l1-regularised estimator for the idiosyncratic VAR matrix, possible values are:
\itemize{
   \item{\code{'lasso'}}{Lasso estimator}
   \item{\code{'ds'}}{Dantzig Selector}
}}

\item{idio.cv.args}{A list specifying arguments to the cross-validation (CV) procedure for the idiosyncratic VAR. This contains:
\itemize{
   \item{\code{'n.folds'}}{number of folds}
   \item{\code{'path.length'}}{number of lambda values to consider}
   \item{\code{'symmetric'}}{symmetrisation method for Gamma matrix}
   \item{\code{'cv.plot'}}{Boolean selecting whether to plot the CV curve}
}}

\item{center}{demean the input `x`}
}
\value{
An S3 object of class \code{fnets}, which contains the following fields:
\itemize{
\item{q: see Arguments}
\item{spec: Spectral density matrices}
\item{acv: Autocovariance matrices}
\item{common.var: Estimated common component}
\item{idio.var: Estimated idiosyncratic component}
\item{mean.x: Removed means of x}
\item{kern.bandwidth.const: Constant to determine bandwidth size}
}
}
\description{
This function estimates the spectral density and autocovariance matrices of the common and the idiosyncratic components, impulse response function and common shocks, and (sparse) VAR transition matrix and innovation covariance matrix.
}
\details{
Further information can be found in Barigozzi, Cho and Owens (2021).
}
\examples{
fnets(sample.data, q=2, idio.method = "lasso")
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
