% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{hl.factor.number}
\alias{hl.factor.number}
\title{Factor number estimator of Hallin and Liska (2011)}
\usage{
hl.factor.number(xx, q.max, mm, w = NULL, do.plot = TRUE)
}
\arguments{
\item{xx}{centred input time series matrix, with each row representing a time series}

\item{q.max}{the maximum number of factors to consider}

\item{mm}{bandwidth}

\item{w}{weights, defaults to Bartlett weights}

\item{do.plot}{return a plot of the information criteria}
}
\value{
A list containing
\itemize{
\item{\code{'q.hat'}}{ Estimated factor numbers corresponding to each criterion}
\item{\code{'Gamma_x'}}{ Autocovariance of x}
\item{\code{'Sigma_x'}}{ Spectral density of x}
\item{\code{'sv'}}{ singular value decomposition of Sigma_x}
}
\examples
    hl.factor.number(sample.data,6, 10)
}
\description{
Selects the factor number `q` based on 6 information criteria
}
\references{
Hallin, M., & Liška, R. (2007). Determining the number of factors in the general dynamic factor model. Journal of the American Statistical Association, 102(478), 603--617.
}
