% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{idio.predict}
\alias{idio.predict}
\title{Prediction for the idiosyncratic VAR process}
\usage{
idio.predict(object, x, cpre, h = 1)
}
\arguments{
\item{object}{\code{fnets} object}

\item{x}{input time series matrix, with each row representing a time series}

\item{cpre}{estimated common component}

\item{h}{forecast horizon}
}
\value{
A list containing
\itemize{
\item{\code{'is'}}{ in-sample estimation}
\item{\code{'fc'}}{ forecast}
\item{\code{'h'}}{ forecast horizon}
}
}
\description{
Predicts idiosyncratic components from a \code{fnets} object for new data
}
\examples{
require(fnets)
model <- fnets(sample.data, q=2, idio.method = "lasso")
pr <- predict(model,sample.data, common.method = "static")
cpre <- common.predict(model,sample.data, common.method = "static")
ip <- idio.predict(model,sample.data, cpre)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
