% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{plot.fnets}
\alias{plot.fnets}
\title{Plot fnets object}
\usage{
\method{plot}{fnets}(object, names = NULL, groups = NULL, threshold = 0, size = NULL, ...)
}
\description{
Plot fnets object
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
