% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{predict.fnets}
\alias{predict.fnets}
\title{Prediction by fnets}
\usage{
\method{predict}{fnets}(object, x, h = 1, common.method = c("static", "var"), r = NULL, ...)
}
\arguments{
\item{object}{\code{fnets} object}

\item{x}{input time series matrix, with each row representing a time series}

\item{h}{forecast horizon}

\item{common.method}{which of "static" or "var" to forecast the common component with}

\item{r}{factor number, if r=NULL this is selected using the maximal eigenratio}

\item{...}{further arguments}
}
\value{
A list containing
\itemize{
\item{\code{'fitted'}}{ \code{x} in-sample estimation}
\item{\code{'forecast'}}{ \code{x} forecast}
\item{\code{'common.pred'}}{ Prediction for the factor-driven common component}
\item{\code{'idio.pred'}}{ Prediction for the idiosyncratic component}
\item{\code{'x.mean'}}{ removed mean of \code{x} }
}
}
\description{
Predicts common and idiosyncratic components from a \code{fnets} object for new data
}
\examples{
require(fnets)
set.seed(222)
n <- 200
p<- 100
chi <- sim.factor.M2(n,p)
xi <- sim.idio(n,p)
sample.data <- chi$data + xi$data
model <- fnets(sample.data, q=2, idio.method = "lasso")
pr <- predict(model,sample.data, common.method = "static")
cpre <- common.predict(model,sample.data, common.method = "static")
ip <- idio.predict(model,sample.data, cpre)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
