% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{predict.fnets}
\alias{predict.fnets}
\title{Prediction by fnets}
\usage{
\method{predict}{fnets}(object, x, h = 1, common.method = c("static", "var"), r = NULL)
}
\arguments{
\item{object}{fnets object}

\item{x}{input time series matrix, with each row representing a time series}

\item{h}{forecast horizon}

\item{common.method}{which of "static" or "var" to forecast the common component with}

\item{r}{factor number, if r=NULL this is selected using the maximal eigenratio}
}
\value{
A list containing
\itemize{
\item{fitted: x in-sample estimation}
\item{forecast: x forecast}
\item{common.pred: Prediction for the factor-driven common component}
\item{idio.pred: Prediction for the idiosyncratic component}
\item{x.mean: removed mean of x}
}
}
\description{
Predicts common and idiosyncratic components from a fnets object for new data
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
