% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{print.fm}
\alias{print.fm}
\title{Print factor model}
\usage{
\method{print}{fm}(x, ...)
}
\arguments{
\item{x}{\code{fm} object}

\item{...}{not used}
}
\value{
NULL, printed to console
}
\description{
Prints a summary of a \code{fm} object
}
\examples{
out <- fnets.factor.model(data.restricted, q = "ic")
print(out)
}
\seealso{
\link[fnets]{fnets.factor.model}
}
