% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnets.R
\name{print.fnets}
\alias{print.fnets}
\title{Print fnets}
\usage{
\method{print}{fnets}(x, ...)
}
\arguments{
\item{x}{\code{fnets} object}

\item{...}{not used}
}
\value{
NULL, printed to console
}
\description{
Prints a summary of a \code{fnets} object
}
\examples{
\donttest{
set.seed(123)
n <- 500
p <- 50
common <- sim.restricted(n, p)
idio <- sim.var(n, p)
x <- common$data + idio$data
out <- fnets(x, q = 2,
do.lrpc = FALSE, var.args = list(n.cores = 2))
print(out)
x <- idio$data
out <- fnets.var(x,
n.cores = 2)
print(out)
}
}
\seealso{
\link[fnets]{fnets}
}
