% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.factor.M1}
\alias{sim.factor.M1}
\title{Simulate data from a dynamic factor model (Model 1)}
\usage{
sim.factor.M1(n, p, q = 2, r = 4, do.scale = T, loadings = NULL, K = NULL)
}
\arguments{
\item{n}{sample size}

\item{p}{number of series}

\item{q}{dynamic dimension (default 2)}

\item{r}{factor number (default 4)}

\item{do.scale}{scale the output (default \code{TRUE} )}

\item{loadings}{loading matrix, dimension p by r (default null)}

\item{K}{transition matrix, dimension r by q (default null)}
}
\value{
A list containing
\itemize{
\item{\code{'data'}}{ generated series}
\item{\code{'shocks'}}{ \code{q}-dimensional shock series}
\item{\code{'factors'}}{ \code{r}-dimensional factor series}
\item{\code{'K'}}{ transition matrix}
\item{\code{'loadings'}}{ factor loadings}
}
}
\description{
Simulate data from a dynamic factor model (Model 1)
}
\examples{
    sim.factor.M1(100,10)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
\seealso{
\link{sim.factor.M2}
}
