% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.factor.M2}
\alias{sim.factor.M2}
\title{Simulate data from a static factor model (Model 2)}
\usage{
sim.factor.M2(
  n,
  p,
  MA_lag = 20,
  do_scale = T,
  a1 = NULL,
  a2 = NULL,
  alpha1 = NULL,
  alpha2 = NULL
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of series}

\item{MA_lag}{lag for MA representation}

\item{do_scale}{scale the ouptut (default true)}

\item{a1, a2, alpha1, alpha2}{generative parameters (default null, see reference)}
}
\value{
A list containing
\itemize{
\item{data: generated series}
\item{shocks: 2-dimensional shock series}
\item{a1,a2,alpha1,alpha2:  generative parameters}
}
}
\description{
Simulate data from a static factor model (Model 2)
}
\examples{
sim.factor.M2(100,10)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
