% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static.pca}
\alias{static.pca}
\title{Static PCA}
\usage{
static.pca(
  xx,
  q = NULL,
  q.method = c("ic", "er"),
  q.max = NULL,
  pen.op = 2,
  kern.bw = NULL,
  mm = NULL
)
}
\description{
Static PCA
}
\keyword{internal}
