% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static.pca}
\alias{static.pca}
\title{Static PCA}
\usage{
static.pca(
  xx,
  q.max = NULL,
  q = NULL,
  q.method = c("bn", "er"),
  ic.op = 2,
  kern.const = 4,
  mm = NULL
)
}
\description{
Static PCA
}
\keyword{internal}
