% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{var.dantzig}
\alias{var.dantzig}
\title{Dantzig selector-type Yule-Walker estimation for VAR processes}
\usage{
var.dantzig(
  GG,
  gg,
  lambda,
  symmetric = "min",
  n.cores = min(parallel::detectCores() - 1, 3)
)
}
\arguments{
\item{GG, gg}{output from `make.gg`}

\item{lambda}{regularisation parameter}

\item{symmetric}{type of symmetry to enforce on `Gamma`, one of 'min', 'max', 'avg', 'none'}

\item{n.cores}{number of cores to use for parallel computing}
}
\value{
A list which contains the following fields:
\itemize{
\item{beta: VAR parameters}
\item{lambda: regularisation parameter}
\item{Gamma: Estimated noise covariance}
}
}
\description{
Returns parameter estimates for the idiosyncratic VAR and the corresponding Gamma matrix
}
\details{
Further information can be found in Barigozzi, Cho and Owens (2021).
}
\examples{
require(fnets)
require(doParallel)
require(lpSolve)
dpca <- dyn.pca(sample.data, q=2)
mg <- make.gg(dpca$acv$Gamma_i,1)
vl <- var.lasso(mg$GG,mg$gg, .1)
vd <- var.dantzig(mg$GG,mg$gg, .1, n.cores = 1)
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
