% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{var.dantzig}
\alias{var.dantzig}
\title{Dantzig selector-type estimator of VAR processes via constrained \code{l1}-minimisation}
\usage{
var.dantzig(
  GG,
  gg,
  lambda,
  symmetric = "min",
  n.cores = min(parallel::detectCores() - 1, 3)
)
}
\description{
Dantzig selector-type estimator of VAR processes via constrained \code{l1}-minimisation
}
\keyword{internal}
