% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{var.lasso}
\alias{var.lasso}
\title{Lasso-type Yule-Walker estimation for VAR processes}
\usage{
var.lasso(
  GG,
  gg,
  lambda,
  symmetric = "min",
  niter = 100,
  tol = 0,
  do.plot = FALSE
)
}
\arguments{
\item{GG, gg}{output from `make.gg`}

\item{lambda}{regularisation parameter}

\item{symmetric}{type of symmetry to enforce on `Gamma`, one of 'min', 'max', 'avg', 'none'}

\item{niter}{maximum number of descent steps}

\item{tol}{numerical tolerance for increases in the loss function}

\item{do.plot}{return a plot of the loss function against descent steps}
}
\value{
A list which contains the following fields:
\itemize{
\item{beta: VAR parameters}
\item{lambda: regularisation parameter}
\item{Gamma: Estimated noise covariance}
\item{loss: Objective function value}
}
}
\description{
Returns parameter estimates for the idiosyncratic VAR and the corresponding Gamma matrix
}
\details{
Further information can be found in Barigozzi, Cho and Owens (2021).
}
\references{
Barigozzi, M., Cho, H., & Owens, D. (2021) Factor-adjusted network analysis for high-dimensional time series.
}
