% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio.R
\name{var.lasso}
\alias{var.lasso}
\title{Lasso-type estimator of VAR processes via \code{l1}-regularised \code{M}-estimation}
\usage{
var.lasso(GG, gg, lambda, symmetric = "min", n.iter = 100, tol = 0)
}
\description{
Lasso-type estimator of VAR processes via \code{l1}-regularised \code{M}-estimation
}
\keyword{internal}
